refactor(nfc): restructure folder layout by domain
