+++ suite/nfc/src/additional-backup/AdditionalBackupSuccess.tsx
@@ -1,10 +1,10 @@
 import { Translation } from '@suite/intl';
-import { Column, H3, Icon, Paragraph } from '@trezor/components';
+import { Column, H3, IconCircle, Paragraph } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
 export const AdditionalBackupSuccess = () => (
     <Column gap={spacings.lg} alignItems="start">
-        <Icon name="checkCircleFilled" size={48} intent="brand" />
+        <IconCircle name="check" size={96} intent="brand" />
         <Column>
             <H3>
                 <Translation id="TR_WALLET_BACKUP_CREATED" />
