+++ suite/nfc/src/index.ts
@@ -1,8 +1,8 @@
-export { AdditionalBackupDisclaimer } from './AdditionalBackupDisclaimer';
-export { AdditionalBackupSteps } from './AdditionalBackupSteps';
-export { AdditionalBackupSuccess } from './AdditionalBackupSuccess';
-export { BackupAlternativeCard } from './BackupAlternativeCard';
-export { CreateNfcBackup } from './CreateNfcBackup';
-export { NfcTag } from './NfcTag';
-export { NoNfcTags } from './NoNfcTags';
-export { SelectBackupType } from './SelectBackupType';
+export { AdditionalBackupDisclaimer } from './additional-backup/AdditionalBackupDisclaimer';
+export { AdditionalBackupSteps } from './additional-backup/AdditionalBackupSteps';
+export { AdditionalBackupSuccess } from './additional-backup/AdditionalBackupSuccess';
+export { BackupAlternativeCard } from './components/BackupAlternativeCard';
+export { CreateNfcBackup } from './create-backup/CreateNfcBackup';
+export { NfcTag } from './components/NfcTag';
+export { NoNfcTags } from './create-backup/NoNfcTags';
+export { SelectBackupType } from './create-backup/SelectBackupType';
